- Refactored code and removed the redundant ProgressingPlayersAndGroups class - Fixed a regression that caused players who were supposed to be eliminated from the tournament to progress to the next stage instead.
